Removed duplication from TestTimestamp.java
